# src/infra/http-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Golang Challenge",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/feira": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "Update feira",
                "parameters": [
                    {
                        "description": "Update feira",
                        "name": "feira",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create feira",
                "parameters": [
                    {
                        "description": "Add feira",
                        "name": "feira",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/feira/bairro/{query}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Find feira by bairro",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feira/distrito/{query}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Find feira by distrito",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feira/nome/{query}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Find feira by nome",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feira/regiao/{query}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Find feira by regiao",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query",
                        "name": "query",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Feira"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/feira/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete feira",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Show server status",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.Feira": {
            "type": "object",
            "properties": {
                "areap": {
                    "type": "integer"
                },
                "bairro": {
                    "type": "string"
                },
                "cod_dist": {
                    "type": "string"
                },
                "cod_sub_pref": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "integer"
                },
                "logradouro": {
                    "type": "string"
                },
                "long": {
                    "type": "integer"
                },
                "nome_feira": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "referencia": {
                    "type": "string"
                },
                "regiao5": {
                    "type": "string"
                },
                "regiao8": {
                    "type": "string"
                },
                "registro": {
                    "type": "string"
                },
                "set_cens": {
                    "type": "integer"
                },
                "sub_prefe": {
                    "type": "string"
                }
            }
        }
    }
}